Replace old mutator test suite with new suite
